Jumpscare and first item pickup added
Deer jumpscare at the house added when a key for the fence is picked up. Added two triggers one for picking up the fence key, second one for the deer jumpscare trigger.
The fence can be opened with the key but the animation restarts it to the first position at the end (needs fixing).