A couple of jumpscares added and map is bigger
Jumpscare from the toilet added.
A square with two houses added and a jumpscare on the second house when you come close to enter it.
Road is longer and has a jumpscare where one of the trees falls behind the back of the player blocking the  old road.
Road leads to church ruins which will be implemented(for now there are no colliders or anything else).