More of the game completed
Road ahead of the church added with two jumpscares,
one after the church gate, second after a small dirty building there is a horror woman in the dumpster that makes a sound when you pass by.
Another big house added with two triggers.
One with a horror woman standing at the edge of the corridor and another with just a baby head playing a creapy laugh. Game slowly coming to the end.